requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
drf-spectacular>=0.27,<1.0
django-filter>=24.2,<25.0
